Add ‘Text.Megaparsec.Internal’
