+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Megaparsec 6.5.0
+
+* Added `Text.Megaparsec.Internal`, which exposes some internal data
+  structures and data constructor of `ParsecT`.
+
 ## Megaparsec 6.4.1
 
 * `scientific` now correctly backtracks after attempting to parse fractional
